+++ src/ui/App.tsx
@@ -1,4 +1,4 @@
-import React from "react"
+import React, { useEffect } from "react"
 import {
     CssBaseline,
     makeStyles,
@@ -6,10 +6,13 @@ import {
     useMediaQuery,
     MuiThemeProvider
 } from "@material-ui/core"
+import { DndProvider } from "react-dnd"
+import HTML5Backend from "react-dnd-html5-backend"
 
 import Navbar from "./components/Navbar"
 import FolderPanel from "./components/FolderPanel"
 import SubfolderPanel from "./components/SubfolderPanel"
+import { useStore } from "./Store"
 
 const useAppStyle = makeStyles({
     container: {
@@ -38,30 +41,39 @@ const useAppStyle = makeStyles({
 })
 
 const App: React.FC = () => {
+    const { darkMode } = useStore()
+
     const classNames = useAppStyle()
 
     const prefersDarkMode = useMediaQuery("(prefers-color-scheme: dark)")
 
+    useEffect(() => {
+        if (darkMode !== prefersDarkMode) {
+        }
+    }, [prefersDarkMode])
+
     const theme = React.useMemo(() => {
         return createMuiTheme({
             palette: {
-                type: prefersDarkMode ? "dark" : "light",
+                type: darkMode ? "dark" : "light",
                 primary: {
                     main: "#3367d6"
                 }
             }
         })
-    }, [prefersDarkMode])
+    }, [darkMode])
 
     return (
         <MuiThemeProvider theme={theme}>
             <div className={classNames.container}>
                 <CssBaseline />
                 <Navbar />
-                <div className={classNames.mainContent}>
-                    <FolderPanel className={classNames.folderPanel} />
-                    <SubfolderPanel className={classNames.subfolderPanel} />
-                </div>
+                <DndProvider backend={HTML5Backend}>
+                    <div className={classNames.mainContent}>
+                        <FolderPanel className={classNames.folderPanel} />
+                        <SubfolderPanel className={classNames.subfolderPanel} />
+                    </div>
+                </DndProvider>
             </div>
         </MuiThemeProvider>
     )